+++ app/src/main/java/woowacourse/movie/domain/SeatGrade.kt
@@ -1,0 +1,9 @@
+package woowacourse.movie.domain
+
+enum class SeatGrade(
+    val price: Int,
+) {
+    S(15_000),
+    A(12_000),
+    B(10_000),
+}
