+++ app/src/test/java/woowacourse/movie/domain/SeatTest.kt
@@ -1,11 +1,8 @@
 package woowacourse.movie.domain
 
 import io.kotest.matchers.throwable.shouldHaveMessage
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.ValueSource
 
 class SeatTest {
     @Test
@@ -28,34 +25,5 @@ class SeatTest {
         // then
         assertThrows<IllegalArgumentException> { Seat(row, column) }
             .shouldHaveMessage("The seats are arranged in 4 columns. But value was $column")
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = [1, 2])
-    fun `1, 2행은 B등급이다`(row: Int) {
-        // when
-        val seat = Seat(row, 1)
-
-        // then
-        assertThat(seat.grade).isEqualTo(SeatGrade.B)
-    }
-
-    @Test
-    fun `B등급 좌석의 가격은 10,000원이다`() {
-        // when
-        val seatGrade = SeatGrade.B
-
-        // then
-        assertThat(seatGrade.price).isEqualTo(10_000)
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = [3, 4])
-    fun `3, 4행은 S등급이다`(row: Int) {
-        // when
-        val seat = Seat(row, 1)
-
-        // then
-        assertThat(seat.grade).isEqualTo(SeatGrade.S)
     }
 }
